fix(admin): reset persisted state across test sessions
